refactor, bugfix
