+++ src/schema.py
@@ -313,39 +313,6 @@ WHERE committee_id={committee_id}
     )
     return query
 
-# fec filings
-
-
-def fec_file_exists(fec_file_id: str) -> SQL:
-    query = sql.SQL('SELECT * FROM fec.filings WHERE fec_file_id={fec_file_id}')\
-        .format(fec_file_id=Literal(fec_file_id))
-    return query
-
-
-def insert_fec_filing(filing: JSONType) -> SQL:
-    values = OrderedDict(sorted(filing.items()))
-
-    query_string = 'INSERT INTO fec.filings ('\
-        + ', '.join([f'{key}' for key, val in values.items()]) + ') '\
-        + 'VALUES (' + ', '.join(['{}' for key, val in values.items()]) + ')'
-
-    query = sql.SQL(query_string)\
-               .format(*[Literal(val) for key, val in values.items()])
-
-    return query
-
-
-def amendment_chain_exists(fec_file_id: str, amendment_id: str) -> SQL:
-    query = sql.SQL('SELECT * FROM fec.filing_amendment_chain WHERE fec_file_id={} AND amendment_id={}')\
-        .format(Literal(fec_file_id), Literal(amendment_id))
-    return query
-
-
-def insert_amendment_chain(fec_file_id: str, amendment_id: str, amendment_number: int) -> SQL:
-    query = sql.SQL('INSERT INTO fec.filing_amendment_chain(fec_file_id, amendment_id, amendment_number) VALUES ({}, {}, {})')\
-        .format(Literal(fec_file_id), Literal(amendment_id), Literal(amendment_number))
-    return query
-
 
 # comittee totals
